+++ docs/changelog.md
@@ -2,6 +2,15 @@
 
 A full changelog of changes, dating all the way back to the first release.
 
+## Version NEXT
+
+Features
+
+- Added a few new variables related to timestamps and time
+  - `%TIMESTAMP%` - replaced with `<t:UNIX_TIMESTAMP>` which Discord will convert to the user's local time
+  - `%TIMESINCE%` - replaced with `<t:UNIX_TIMESTAMP:R>` which Discord will convert to a relative time (e.g. 2 hours ago)
+  - `%UNIX_TIMESTAMP%` - replaced with the UNIX timestamp of the event (e.g. 1634567890). This can be used to create a custom timestamp format in the message.
+
 ## Version 2.3.0
 
 Features